Crate plexus

Plexus is a highly composable library for polygonal mesh processing.

Versions of Plexus in the 0.0.* series are experimental and unstable. Consider depending on the development branch of the repository. See the website for the latest information and documentation.

Modules

buffer

Linear representations of polygonal meshes.

builder

Incremental polygonal mesh construction.

constant

Morphisms between constant generics and numeric types.

encoding

Serialization and encodings.

geometry

Geometric traits and computational geometry.

graph

Half-edge graph representation of polygonal meshes.

index

Indexing and aggregation.

integration

Integration of external crates and foreign types.

prelude

Re-exports of commonly used types and traits.

primitive

Primitive topological structures.

Structs

Keys

Iterator that maps graph views to their keys.

Perimeter

Iterator that produces a window of duplets over its input.

Enums

MeshArity

Arity of a compound structure.

Traits

Arity

Arity of primitives and polygonal meshes.

DynamicArity

Value-level arity.

IteratorExt

Extension methods for types implementing Iterator.

Monomorphic

Topological types with fixed and singular arity.

StaticArity

Type-level arity.

TryFromIterator