Struct plexus::buffer::MeshBuffer

pub struct MeshBuffer<R, G> where
    R: Grouping,  { /* fields omitted */ }

Polygonal mesh composed of vertex and index buffers.

A MeshBuffer is a linear representation of a polygonal mesh that is composed of two separate but related linear buffers: an index buffer and a vertex buffer. The index buffer contains ordered indices into the data in the vertex buffer and describes the topology of the mesh. The vertex buffer contains arbitrary data that describes each vertex.

MeshBuffer only explicitly respresents vertices via the vertex buffer and surfaces via the index buffer. There is no explicit representation of structures like edges and faces.

The R type parameter specifies the Grouping of the index buffer. See the index module documention for more information.

Implementations

impl<R, G> MeshBuffer<R, G> where
R: Grouping,
Vec<R::Group>: IndexBuffer<R>,

pub fn new() -> Self

Creates an empty MeshBuffer.

Examples

use plexus::buffer::MeshBuffer;
use plexus::index::Flat3;

let buffer = MeshBuffer::<Flat3<u64>, (f64, f64, f64)>::new();

impl<R, G> MeshBuffer<R, G> where
R: Grouping,

pub fn into_raw_buffers(self) -> (Vec<R::Group>, Vec<G>)

Converts a MeshBuffer into its index and vertex buffers.

pub fn map_vertices<H, F>(self, f: F) -> MeshBuffer<R, H> where
F: FnMut(G) -> H,

Maps over the vertex data in a MeshBuffer.

Examples

Translating the position data in a buffer:

use decorum::N64;
use nalgebra::{Point3, Vector3};
use plexus::buffer::MeshBuffer3;
use plexus::prelude::*;
use plexus::primitive::generate::Position;
use plexus::primitive::sphere::UvSphere;

let buffer: MeshBuffer3<usize, Point3<f64>> = UvSphere::new(16, 8)
    .polygons::<Position<Point3<N64>>>()
    .triangulate()
    .collect();
// Translate the positions.
let translation = Vector3::<f64>::x() * 2.0;
let buffer = buffer.map_vertices(|position| position + translation);

pub fn as_index_slice(&self) -> &[R::Group]

Gets a slice over the index data.

pub fn as_vertex_slice(&self) -> &[G]

Gets a slice over the vertex data.

impl<T, G, const N: usize> MeshBuffer<Flat<T, N>, G> where
Constant<N>: ToType,
TypeOf<N>: NonZero,
T: Copy + Integer + NumCast + Unsigned,

pub fn append<R, H>(
&mut self,
buffer: &mut MeshBuffer<R, H>
) -> Result<(), BufferError> where
G: FromGeometry<H>,
R: Grouping,
R::Group: Into<<Flat<T, N> as Grouping>::Group>,

Appends the contents of a flat MeshBuffer into another MeshBuffer. The source buffer is drained.

Errors

Returns an error if an index overflows.

impl<P, G> MeshBuffer<P, G> where
P: Grouping + Polygonal,
P::Vertex: Copy + Integer + NumCast + Unsigned,

pub fn append<R, H>(
&mut self,
buffer: &mut MeshBuffer<R, H>
) -> Result<(), BufferError> where
G: FromGeometry<H>,
R: Grouping,
R::Group: Into<<P as Grouping>::Group>,
<P as Grouping>::Group: Map<P::Vertex, Output = <P as Grouping>::Group> + Topological<Vertex = P::Vertex>,

Appends the contents of a structured MeshBuffer into another MeshBuffer. The source buffer is drained.

Errors

Returns an error if an index overflows.