Trait plexus::buffer::FromRawBuffersWithArity

pub trait FromRawBuffersWithArity<N, G>: Sized {
    type Error: Debug;
    fn from_raw_buffers_with_arity<I, J>(
        indices: I, 
        vertices: J, 
        arity: usize
    ) -> Result<Self, Self::Error>
    where
        I: IntoIterator<Item = N>,
        J: IntoIterator<Item = G>;
}

Conversion from raw buffers that do not encode their arity.

Associated Types

type Error: Debug

Loading content...
Required methods

fn from_raw_buffers_with_arity<I, J>(
indices: I,
vertices: J,
arity: usize
) -> Result<Self, Self::Error> where
I: IntoIterator<Item = N>,
J: IntoIterator<Item = G>,

Creates a type from raw buffers with the given arity.

Errors

Returns an error if the raw buffers are inconsistent, the raw buffers are incompatible with the given arity, or the implementor cannot be constructed from the buffers. The latter typically occurs if the given topology is unsupported.

Loading content...