Trait plexus::builder::Buildable

pub trait Buildable: Sized {
    type Builder: MeshBuilder<Commit = Self, Error = Self::Error, Vertex = Self::Vertex, Facet = Self::Facet>;
    type Error: Debug;
    type Vertex;
    type Facet: Default;
    fn builder() -> Self::Builder;
}

Polygonal mesh data structure that can be built incrementally.

This trait is the primary entrypoint into the builder API. Types that implement this trait expose a MeshBuilder that can be used to construct an instance of the type from surfaces and facets.

Associated Types

type Builder: MeshBuilder<Commit = Self, Error = Self::Error, Vertex = Self::Vertex, Facet = Self::Facet>

type Error: Debug

type Vertex

Vertex data.

This type represents the data associated with vertices in the mesh. This typically includes positional data, but no data is required and this type may be the unit type ().

Each builder trait also exposes such an associated type which is constrained by the Builder type.

type Facet: Default

Facet data.

This type represents the data associated with facets in the mesh. No data is required and this type may be the unit type ().

Each builder trait also exposes such an associated type which is constrained by the Builder type.

Loading content...
Required methods

fn builder() -> Self::Builder

Loading content...