Module plexus::geometry

Geometric traits and computational geometry.

Plexus uses the theon crate to abstract over types that represent Euclidean spaces and implement linear algebra. Types and traits are re-exported from theon in this module, but it may be necessary to import additional types from theon. See the integration module.

Modules

partition

Structs

Aabb
Axis-aligned bounding box.

Line
Line.

Plane

Ray
Ray or half-line.

Unit
Unit vector.

Enums

LineLine
Intersection of lines.

LinePlane
Intersection of a line and a plane.

PlaneRay
Intersection of a plane and a ray.

Traits

AsPosition
Immutable positional data.

AsPositionMut
Mutable positional data.

FromGeometry

Intersection
Intersection of geometric objects.

IntoGeometry

Metric

UnitGeometry
Geometry elision.

Type Definitions

Position

Scalar
The scalar of a EuclideanSpace.

Vector
The vector (translation, coordinate space, etc.) of a EuclideanSpace.