Struct plexus::graph::Ring

pub struct Ring<B> where
    B: Reborrow,
    B::Target: AsStorage<Arc<<B as Parametric>::Data>> + Consistent + Parametric,  { /* fields omitted */ }

Closed path formed by adjacent arcs.

Rings are closed paths formed by arcs and their immediately adjacent arcs. In a consistent graph, every arc forms such a path. Such paths may or may not be occupied by faces.

Rings are notated by their path. A ring with a perimeter formed by vertices $A$, $B$, and $C$ is notated $\overrightarrow{\{A,B,C\}}$. Note that rotations of the set of vertices are equivalent, such as $\overrightarrow{\{B,C,A\}}$.

Implementations

impl<B, M, G> Ring<B> where
B: Reborrow<Target = M>,
M: AsStorage<Arc<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn to_ref(&self) -> Ring<&M>

impl<'a, B, M, G> Ring<B> where
B: ReborrowInto<'a, Target = M>,
M: AsStorage<Arc<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn into_ref(self) -> Ring<&'a M>

impl<B, M, G> Ring<B> where
B: Reborrow<Target = M>,
M: AsStorage<Arc<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn into_arc(self) -> ArcView<B>

Converts the ring into its leading arc.

pub fn arc(&self) -> ArcView<&M>

Gets the leading arc of the ring.

impl<B, M, G> Ring<B> where
B: Reborrow<Target = M>,
M: AsStorage<Arc<G>> + AsStorage<Vertex<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn into_path(self) -> Path<'static, B>

pub fn path(&self) -> Path<'static, &M>

pub fn shortest_metric(
&self,
from: Selector<VertexKey>,
to: Selector<VertexKey>
) -> Result<usize, GraphError>

Gets the shortest logical metric between vertices within the ring.

The logical metric assigns the unit weight (one) to every arc and effectively counts the number of arcs between vertices. In a closed path like a ring, there are two sub-paths between any two vertices. This function computes the metric of the shortest sub-path. This may be the null path with a zero metric.

impl<B, M, G> Ring<B> where
B: Reborrow<Target = M>,
M: AsStorage<Arc<G>> + AsStorage<Face<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn into_face(self) -> Option<FaceView<B>>

Converts the ring into its face.

If the path has no associated face, then None is returned.

pub fn face(&self) -> Option<FaceView<&M>>

Gets the face of the ring.

If the path has no associated face, then None is returned.

impl<'a, B, M, G> Ring<B> where
B: ReborrowInto<'a, Target = M>,
M: 'a + AsStorage<Arc<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn into_arcs(self) -> impl Clone + Iterator<Item = ArcView<&'a M>>

impl<B, G> Ring<B> where
B: Reborrow,
B::Target: AsStorage<Arc<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn arcs(&self) -> impl Clone + Iterator<Item = ArcView<&B::Target>>

Gets an iterator of views over the arcs within the ring.

impl<'a, B, M, G> Ring<B> where
B: ReborrowInto<'a, Target = M>,
M: 'a + AsStorage<Arc<G>> + AsStorage<Vertex<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn into_vertices(self) -> impl Clone + Iterator<Item = VertexView<&'a M>>

impl<B, G> Ring<B> where
B: Reborrow,
B::Target: AsStorage<Arc<G>> + AsStorage<Vertex<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn vertices(&self) -> impl Clone + Iterator<Item = VertexView<&B::Target>>

Gets an iterator of views over the vertices within the ring.

impl<'a, M, G> Ring<&'a mut M> where
M: AsStorageMut<Arc<G>> + Consistent + Parametric<Data = G>,
G: 'a + GraphData,

pub fn into_arc_orphans(self) -> impl Iterator<Item = ArcOrphan<'a, G>>

impl<B> Ring<B> where
B: ReborrowMut,
B::Target: AsStorageMut<Arc<<B as Parametric>::Data>> + Consistent + Parametric,

pub fn arc_orphans(
&mut self
) -> impl Iterator<Item = ArcOrphan<'_, <B as Parametric>::Data>>

Gets an iterator of orphan views over the arcs in the ring.

impl<'a, M, G> Ring<&'a mut M> where
M: AsStorage<Arc<G>> + AsStorageMut<Vertex<G>> + Consistent + Parametric<Data = G>,
G: 'a + GraphData,

pub fn into_vertex_orphans(self) -> impl Iterator<Item = VertexOrphan<'a, G>>

impl<B> Ring<B> where
B: ReborrowMut,
B::Target: AsStorage<Arc<<B as Parametric>::Data>> + AsStorageMut<Vertex<<B as Parametric>::Data>> + Consistent + Parametric,

pub fn vertex_orphans(
&mut self
) -> impl Iterator<Item = VertexOrphan<'_, <B as Parametric>::Data>>

Gets an iterator of views over the vertices within the ring.

impl<'a, M, G> Ring<&'a mut M> where
M: AsStorage<Vertex<G>> + AsStorage<Arc<G>> + AsStorage<Face<G>> + Default + Mutable<Data = G>,
G: GraphData,

pub fn get_or_insert_face(self) -> FaceView<&'a mut M>

Gets the face of the ring or inserts a face if one does not already exist.

Returns the existing or inserted face.

pub fn get_or_insert_face_with<F>(self, f: F) -> FaceView<&'a mut M> where
F: FnOnce() -> G::Face,

Gets the face of the ring or inserts a face if one does not already exist.

If a face is inserted, then the given function is used to get the data for the face.

Returns the existing or inserted face.