Enum plexus::primitive::BoundedPolygon

pub enum BoundedPolygon<G> {
    N3(Trigon<G>),
    N4(Tetragon<G>),
}

Bounded polymorphic $n$-gon.

BoundedPolygon represents an $n$-gon with at least three edges by enumerating NGons. As such, $n$ is bounded, because the enumeration only supports a limited set of polygons. Only common arities used by generators are provided.

See the module documentation for more information.

Variants

N3(Trigon<G>)

N4(Tetragon<G>)