Struct plexus::primitive::generate::Normal

pub struct Normal<S = ()> { /* fields omitted */ }

Meta-attribute for surface normals.

Describes the surface normals of a polytope. The generated data is derived from the type parameter S, which typically requires EuclideanSpace.

Examples

Generating raw buffers with normal data of a $uv$-sphere:

use decorum::R64;
use nalgebra::Point3;
use plexus::index::{Flat3, HashIndexer};
use plexus::prelude::*;
use plexus::primitive::generate::Normal;
use plexus::primitive::sphere::UvSphere;

let (indices, normals) = UvSphere::new(8, 8)
    .polygons::<Normal<Point3<R64>>>()
    .map_vertices(|normal| normal.into_inner())
    .triangulate()
    .index_vertices::<Flat3, _>(HashIndexer::default());