Trait plexus::primitive::generate::Attribute

pub trait Attribute { }

Geometric attribute.

Types implementing this trait can be used with Generator to query geometric attributes. For example, the Position type can be used to get positional data for cubes or spheres via Cube and UvSphere.