Struct plexus::primitive::NGon

pub struct NGon<G, const N: usize>(pub [G; N]);

Monomorphic $n$-gon.

NGon represents a polygonal structure as an array. Each array element represents vertex data in order with adjacent elements being connected by an implicit undirected edge. For example, an NGon with three vertices (NGon<_, 3>) would represent a trigon (triangle). Generally these elements are labeled $A$, $B$, $C$, etc. Note that the constant parameter N represents the number of the NGon’s vertices and not the number of its edges (arity).

NGons with less than three vertices are a degenerate case. An NGon with two vertices (NGon<_, 2>) is considered a monogon and is used to represent edges. Such an NGon is not considered a digon, as it represents a single undirected edge rather than two distinct (but collapsed) edges. Note that the polygonal types BoundedPolygon and UnboundedPolygon never represent edges. See the Edge type definition.

NGons with one or zero vertices are unsupported and lack various trait implementations.

See the module documentation for more information.

Tuple Fields
0: [G; N]
Implementations

impl<G, const N: usize> NGon<G, N>

pub fn into_array(self) -> [G; N]

pub fn positions(&self) -> NGon<&Position<G>, N> where
G: AsPosition,

impl<'a, G, const N: usize> NGon<&'a G, N>

pub fn cloned(self) -> NGon<G, N> where
G: Clone,

impl<G> NGon<G, 2_usize>

pub fn new(a: G, b: G) -> Self

pub fn line(&self) -> Option<Line<Position<G>>> where
G: AsPosition,

pub fn is_bisected(&self, other: &Self) -> bool where
G: AsPosition,
Position<G>: FiniteDimensional<N = U2>,

impl<G> NGon<G, 3_usize>

pub fn new(a: G, b: G, c: G) -> Self

pub fn plane(&self) -> Option<Plane<Position<G>>> where
G: AsPosition,
Position<G>: EuclideanSpace + FiniteDimensional<N = U3>,
Vector<Position<G>>: Cross<Output = Vector<Position<G>>>,

impl<G> NGon<G, 4_usize>

pub fn new(a: G, b: G, c: G, d: G) -> Self