Struct plexus::primitive::UnboundedPolygon

pub struct UnboundedPolygon<G>(_);

Unbounded polymorphic $n$-gon.

UnboundedPolygon represents an $n$-gon with three or more edges. Unlike BoundedPolygon, there is no limit to the arity of polygons that UnboundedPolygon may represent.

It is not possible to trivially convert an UnboundedPolygon into other topological types like NGon. See the module documentation for more information.

Implementations

impl<G> UnboundedPolygon<G>

pub fn trigon(a: G, b: G, c: G) -> Self

pub fn tetragon(a: G, b: G, c: G, d: G) -> Self

pub fn positions(&self) -> UnboundedPolygon<&Position<G>> where
G: AsPosition,

impl<'a, G> UnboundedPolygon<&'a G> where
G: Clone,

pub fn cloned(self) -> UnboundedPolygon<G>