Trait plexus::DynamicArity

pub trait DynamicArity: StaticArity {
    type Dynamic: Arity;
    fn arity(&self) -> Self::Dynamic;
}

Value-level arity.

This trait specifies the arity of a value at runtime. This is often distinct from the type-level arity of the StaticArity trait, which expresses the capabilities of a type.

Associated Types

type Dynamic: Arity

Loading content...
Required methods

fn arity(&self) -> Self::Dynamic

Loading content...