Trait plexus::Monomorphic

pub trait Monomorphic: StaticArity<Static = usize> { }

Topological types with fixed and singular arity.

Types are monomorphic if they have a fixed and singular arity as types and values. For example, Trigon always and only represents a trigon (triangle) with an arity of three. Trigon values always have an arity of three and types composed of only Trigons have a compound arity of three.

This contrasts polymorphic types like BoundedPolygon, which have an interval arity at the type-level and a singular but varying arity for values (because a BoundedPolygon value may be either a trigon or tertragon).