Trait plexus::StaticArity

pub trait StaticArity {
    type Static: Arity;

    const ARITY: Self::Static;
}

Type-level arity.

This trait specifies the arity that a type supports. Values of a StaticArity type have an arity that reflects this constant, which may be any type or form implementing the Arity trait.

Associated Types

type Static: Arity

Loading content...
Associated Constants

const ARITY: Self::Static

Loading content...