Trait plexus::buffer::FromRawBuffers

pub trait FromRawBuffers<N, G>: Sized {
    type Error: Debug;

    fn from_raw_buffers<I, J>(
        indices: I, 
        vertices: J
    ) -> Result<Self, Self::Error>
    where
        I: IntoIterator<Item = N>,
        J: IntoIterator<Item = G>;
}

Conversion from raw buffers.

Required Associated Types

type Error: Debug

Required Methods

fn from_raw_buffers<I, J>(indices: I, vertices: J) -> Result<Self, Self::Error> where
I: IntoIterator<Item = N>,
J: IntoIterator<Item = G>,

Creates a type from raw buffers.

Errors

Returns an error if the raw buffers are inconsistent or the implementor cannot be constructed from the buffers. The latter typically occurs if the given topology is unsupported.