Trait plexus::builder::FacetBuilder

pub trait FacetBuilder<K>: ClosedInput where
    K: Eq + Hash, {
    type Key: Copy + Eq + Hash;
    type Facet: Default;

    fn insert_facet<T, U>(
        &mut self, 
        keys: T, 
        data: U
    ) -> Result<Self::Key, Self::Error>
    where
        Self::Facet: FromGeometry<U>,
        T: AsRef<[K]>;
}

Required Associated Types

type Key: Copy + Eq + Hash

Facet key.

Each facet is associated with a key of this type.

type Facet: Default

Required Methods

fn insert_facet<T, U>(
&mut self,
keys: T,
data: U
) -> Result<Self::Key, Self::Error> where
Self::Facet: FromGeometry<U>,
T: AsRef<[K]>,

Inserts a facet into the associated surface.

A facet is formed from connectivity between vertices represented by an ordered slice of vertex keys from the associated SurfaceBuilder.

Returns a key that refers to the inserted facet.