Module plexus::encoding

Serialization and encodings.

This module provides encoding support enabled via Cargo features. Each enabled encoding has a corresponding sub-module. For example, when PLY support is enabled, the ply module is exposed. The following table summarizes the encodings supported by Plexus:

Feature	Default	Encoding	Read	Write

encoding-ply	No	PLY	Yes	No

This module provides traits used by all encodings. These traits describe the outputs and inputs of decoders and encoders, respectively. Generally, these traits should not be used directly. Instead, prefer the conversion traits exposed for specific encodings, such as FromPly when using PLY.

Modules

ply

PLY encoding.

Traits

FaceDecoder

FromEncoding

VertexDecoder