Trait plexus::geometry::UnitGeometry

pub trait UnitGeometry { }

Geometry elision.

Geometric types that implement this trait may be elided. In particular, these types may be converted into and from () via the FromGeometry and IntoGeometry traits.

For a geometric type T, the following table illustrates the elisions in which T may participate:

Bounds on T	From	Into

UnitGeometry	T	()

Default + UnitGeometry	()	T

These conversions are useful when converting between mesh data structures with incompatible geometry, such as from a MeshGraph with face geometry to a MeshBuffer that cannot support such geometry.

When geometry features are enabled, UnitGeometry is implemented for integrated foreign types.