Struct plexus::graph::FaceView

pub struct FaceView<B> where
    B: Reborrow,
    B::Target: AsStorage<Face<<B as Parametric>::Data>> + Parametric,  { /* private fields */ }

View of a face entity.

Faces are notated by the path of their associated ring. A triangular face with a perimeter formed by vertices $A$, $B$, and $C$ is notated $\overrightarrow{\{A,B,C\}}$. While the precise ordering of vertices is determined by a face’s leading arc, the same face may be notated using rotations of this set, such as $\overrightarrow{\{B,C,A\}}$.

See the graph module documentation for more information about views.

Implementations

impl<B, M> FaceView<B> where
B: Reborrow<Target = M>,
M: AsStorage<Face<<B as Parametric>::Data>> + Parametric,

pub fn to_ref(&self) -> FaceView<&M>

impl<'a, B, M, G> FaceView<B> where
B: ReborrowInto<'a, Target = M>,
M: AsStorage<Face<G>> + Parametric<Data = G>,
G: GraphData,

pub fn into_ref(self) -> FaceView<&'a M>

Examples

use decorum::R64;
use nalgebra::Point3;
use plexus::graph::MeshGraph;
use plexus::prelude::*;
use plexus::primitive::cube::Cube;
use plexus::primitive::generate::Position;

type E3 = Point3<R64>;

let mut graph: MeshGraph<E3> = Cube::new().polygons::<Position<E3>>().collect();
let key = graph.faces().nth(0).unwrap().key();
let face = graph
    .face_mut(key)
    .unwrap()
    .extrude_with_offset(1.0)
    .unwrap()
    .into_ref();

// This would not be possible without conversion into an immutable view.
let _ = face.into_arc();
let _ = face.into_arc().into_next_arc();

impl<B, M, G> FaceView<B> where
B: Reborrow<Target = M>,
M: AsStorage<Face<G>> + Parametric<Data = G>,
G: GraphData,

pub fn get<'a>(&'a self) -> &'a G::Face where
G: 'a,

impl<B, M, G> FaceView<B> where
B: ReborrowMut<Target = M>,
M: AsStorageMut<Face<G>> + Parametric<Data = G>,
G: GraphData,

pub fn get_mut<'a>(&'a mut self) -> &'a mut G::Face where
G: 'a,

impl<B, M, G> FaceView<B> where
B: Reborrow<Target = M>,
M: AsStorage<Arc<G>> + AsStorage<Face<G>> + Parametric<Data = G>,
G: GraphData,

This impl block contains no items.

Reachable API.

impl<B, M, G> FaceView<B> where
B: Reborrow<Target = M>,
M: AsStorage<Arc<G>> + AsStorage<Face<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn into_ring(self) -> Ring<B>

Converts the face into its ring.

pub fn into_arc(self) -> ArcView<B>

Converts the face into its leading arc.

pub fn ring(&self) -> Ring<&M>

Gets the ring of the face.

pub fn arc(&self) -> ArcView<&M>

Gets the leading arc of the face.

impl<B, M, G> FaceView<B> where
B: Reborrow<Target = M>,
M: AsStorage<Arc<G>> + AsStorage<Face<G>> + AsStorage<Vertex<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn centroid(&self) -> VertexPosition<G> where
G: FaceCentroid,
G::Vertex: AsPosition,

pub fn normal(&self) -> Result<Vector<VertexPosition<G>>, GraphError> where
G: FaceNormal,
G::Vertex: AsPosition,

pub fn plane(&self) -> Result<Plane<VertexPosition<G>>, GraphError> where
G: FacePlane,
G::Vertex: AsPosition,
VertexPosition<G>: FiniteDimensional<N = U3>,

impl<B, M, G> FaceView<B> where
B: ReborrowMut<Target = M>,
M: AsStorage<Arc<G>> + AsStorage<Face<G>> + AsStorageMut<Vertex<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn flatten(&mut self) -> Result<(), GraphError> where
G: FacePlane,
G::Vertex: AsPositionMut,
VertexPosition<G>: EuclideanSpace + FiniteDimensional<N = U3>,

Flattens the face by translating the positions of all vertices into a best-fit plane.

Errors

Returns an error if a best-fit plane could not be computed or positions could not be translated into the plane.

impl<'a, B, M, G> FaceView<B> where
B: ReborrowInto<'a, Target = M>,
M: 'a + AsStorage<Arc<G>> + AsStorage<Face<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn into_adjacent_arcs(self) -> impl Clone + Iterator<Item = ArcView<&'a M>>

pub fn into_adjacent_faces(
self
) -> impl Clone + Iterator<Item = FaceView<&'a M>>

impl<B, G> FaceView<B> where
B: Reborrow,
B::Target: AsStorage<Arc<G>> + AsStorage<Face<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn adjacent_arcs(&self) -> impl Clone + Iterator<Item = ArcView<&B::Target>>

Gets an iterator of views over the arcs in the face’s ring.

pub fn adjacent_faces(
&self
) -> impl Clone + Iterator<Item = FaceView<&B::Target>>

Gets an iterator of views over adjacent faces.

impl<'a, B, M, G> FaceView<B> where
B: ReborrowInto<'a, Target = M>,
M: 'a + AsStorage<Arc<G>> + AsStorage<Face<G>> + AsStorage<Vertex<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn into_adjacent_vertices(
self
) -> impl Clone + Iterator<Item = VertexView<&'a M>>

impl<B, G> FaceView<B> where
B: Reborrow,
B::Target: AsStorage<Arc<G>> + AsStorage<Face<G>> + AsStorage<Vertex<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn adjacent_vertices(
&self
) -> impl Clone + Iterator<Item = VertexView<&B::Target>>

Gets an iterator of views over the vertices that form the face.

impl<'a, M, G> FaceView<&'a mut M> where
M: AsStorageMut<Arc<G>> + AsStorage<Face<G>> + Consistent + Parametric<Data = G>,
G: 'a + GraphData,

pub fn into_adjacent_arc_orphans(self) -> impl Iterator<Item = ArcOrphan<'a, G>>

impl<B> FaceView<B> where
B: ReborrowMut,
B::Target: AsStorageMut<Arc<<B as Parametric>::Data>> + AsStorage<Face<<B as Parametric>::Data>> + Consistent + Parametric,

pub fn adjacent_arc_orphans(
&mut self
) -> impl Iterator<Item = ArcOrphan<'_, <B as Parametric>::Data>>

Gets an iterator of orphan views over the arcs in the face’s ring.

impl<'a, M, G> FaceView<&'a mut M> where
M: AsStorage<Arc<G>> + AsStorageMut<Face<G>> + Consistent + Parametric<Data = G>,
G: 'a + GraphData,

pub fn into_adjacent_face_orphans(
self
) -> impl Iterator<Item = FaceOrphan<'a, G>>

impl<B> FaceView<B> where
B: ReborrowMut,
B::Target: AsStorage<Arc<<B as Parametric>::Data>> + AsStorageMut<Face<<B as Parametric>::Data>> + Consistent + Parametric,

pub fn adjacent_face_orphans(
&mut self
) -> impl Iterator<Item = FaceOrphan<'_, <B as Parametric>::Data>>

Gets an iterator of orphan views over adjacent faces.

impl<'a, M> FaceView<&'a mut M> where
M: AsStorage<Arc<<M as Parametric>::Data>> + AsStorage<Face<<M as Parametric>::Data>> + AsStorageMut<Vertex<<M as Parametric>::Data>> + Consistent + Parametric,

pub fn into_adjacent_vertex_orphans(
self
) -> impl Iterator<Item = VertexOrphan<'a, <M as Parametric>::Data>>

impl<B> FaceView<B> where
B: ReborrowMut,
B::Target: AsStorage<Arc<<B as Parametric>::Data>> + AsStorage<Face<<B as Parametric>::Data>> + AsStorageMut<Vertex<<B as Parametric>::Data>> + Consistent + Parametric,

pub fn adjacent_vertex_orphans(
&mut self
) -> impl Iterator<Item = VertexOrphan<'_, <B as Parametric>::Data>>

Gets an iterator of orphan views over the vertices that form the face.

impl<B, G> FaceView<B> where
B: Reborrow,
B::Target: AsStorage<Arc<G>> + AsStorage<Face<G>> + Consistent + Parametric<Data = G>,
G: GraphData,

pub fn traverse_by_breadth(
&self
) -> impl Clone + Iterator<Item = FaceView<&B::Target>>

Gets an iterator that traverses adjacent faces by breadth.

The traversal moves from the face to its adjacent faces and so on. If there are disjoint sub-graphs in the graph, then a traversal will not reach every face in the graph.

pub fn traverse_by_depth(
&self
) -> impl Clone + Iterator<Item = FaceView<&B::Target>>

Gets an iterator that traverses adjacent faces by depth.

The traversal moves from the face to its adjacent faces and so on. If there are disjoint sub-graphs in the graph, then a traversal will not reach every face in the graph.

impl<'a, M, G> FaceView<&'a mut M> where
M: AsStorage<Arc<G>> + AsStorage<Edge<G>> + AsStorage<Face<G>> + AsStorage<Vertex<G>> + Default + Mutable<Data = G>,
G: GraphData,

pub fn split(
self,
source: impl Into<Selector<VertexKey>>,
destination: impl Into<Selector<VertexKey>>
) -> Result<ArcView<&'a mut M>, GraphError>

Splits the face by bisecting it with a composite edge inserted between two non-adjacent vertices within the face’s perimeter.

The vertices can be chosen by key or index, where index selects the $n^\text{th}$ vertex within the face’s ring.

Returns the arc inserted from the source vertex to the destination vertex. If a face $\overrightarrow{\{A,B, C,D\}}$ is split from $A$ to $C$, then it will be decomposed into faces in the rings $\overrightarrow{\{A,B,C\}}$ and $\overrightarrow{\{C,D,A\}}$ and the arc $\overrightarrow{AC}$ will be returned.

Errors

Returns an error if either of the given vertices cannot be found, are not within the face’s perimeter, or the distance between the vertices along the ring is less than two.

Examples

Splitting a quadrilateral face:

use nalgebra::Point2;
use plexus::graph::MeshGraph;
use plexus::prelude::*;
use plexus::primitive::Tetragon;

let mut graph = MeshGraph::<Point2<f64>>::from_raw_buffers(
    vec![Tetragon::new(0usize, 1, 2, 3)],
    vec![(0.0, 0.0), (1.0, 0.0), (1.0, 1.0), (0.0, 1.0)],
)
.unwrap();
let key = graph.faces().nth(0).unwrap().key();
let arc = graph
    .face_mut(key)
    .unwrap()
    .split(ByIndex(0), ByIndex(2))
    .unwrap()
    .into_ref();

pub fn merge(
self,
destination: impl Into<Selector<FaceKey>>
) -> Result<Self, GraphError>

Merges the face into an adjacent face over a shared edge.

The adjacent face can be chosen by key or index, where index selects the $n^\text{th}$ adjacent face.

Returns the merged face.

Errors

Returns an error if the destination face cannot be found or is not adjacent to the initiating face.

Examples

Merging two adjacent quadrilateral faces:

use nalgebra::Point2;
use plexus::graph::MeshGraph;
use plexus::prelude::*;
use plexus::primitive::Tetragon;

let mut graph = MeshGraph::<Point2<f64>>::from_raw_buffers(
    vec![Tetragon::new(0usize, 1, 2, 3), Tetragon::new(0, 3, 4, 5)],
    vec![
        (0.0, 0.0),  // 0
        (1.0, 0.0),  // 1
        (1.0, 1.0),  // 2
        (0.0, 1.0),  // 3
        (-1.0, 1.0), // 4
        (-1.0, 0.0), // 5
    ],
)
.unwrap();

let key = graph.faces().nth(0).unwrap().key();
let face = graph
    .face_mut(key)
    .unwrap()
    .merge(ByIndex(0))
    .unwrap()
    .into_ref();

pub fn bridge(self, destination: FaceKey) -> Result<(), GraphError>

Connects faces with equal arity with faces inserted along their perimeters.

The inserted faces are always quadrilateral. Both the initiating face and destination face are removed.

Errors

Returns an error if the destination face cannot be found or the arity of the face and its destination are not the same.

pub fn triangulate(self) -> Self

Decomposes the face into triangles. Does nothing if the face is triangular.

Returns the terminating face of the decomposition.

pub fn poke_with<F>(self, f: F) -> VertexView<&'a mut M> where
F: FnOnce() -> G::Vertex,

Subdivides the face about a vertex. A triangle fan is formed from each arc in the face’s perimeter and the vertex.

Poking inserts a new vertex with data provided by the given function.

Returns the inserted vertex.

Examples

Forming a pyramid from a triangular face:

use nalgebra::Point3;
use plexus::geometry::{AsPosition, AsPositionMut};
use plexus::graph::MeshGraph;
use plexus::prelude::*;
use plexus::primitive::Trigon;

let mut graph = MeshGraph::<Point3<f64>>::from_raw_buffers(
    vec![Trigon::new(0usize, 1, 2)],
    vec![(-1.0, 0.0, 0.0), (1.0, 0.0, 0.0), (0.0, 2.0, 0.0)],
)
.unwrap();
let key = graph.faces().nth(0).unwrap().key();
let mut face = graph.face_mut(key).unwrap();

// See also `poke_with_offset`, which provides this functionality.
let position = face.centroid() + face.normal().unwrap();
face.poke_with(move || position);

pub fn poke_at_centroid(self) -> VertexView<&'a mut M> where
G: FaceCentroid,
G::Vertex: AsPositionMut,

Subdivides the face about its centroid. A triangle fan is formed from each arc in the face’s perimeter and a vertex inserted at the centroid.

Returns the inserted vertex.

pub fn poke_with_offset<T>(
self,
offset: T
) -> Result<VertexView<&'a mut M>, GraphError> where
T: Into<Scalar<VertexPosition<G>>>,
G: FaceCentroid + FaceNormal,
G::Vertex: AsPositionMut,
VertexPosition<G>: EuclideanSpace,

Subdivides the face about its centroid. A triangle fan is formed from each arc in the face’s perimeter and a vertex inserted at the centroid. The inserted vertex is then translated along the initiating face’s normal by the given offset.

Returns the inserted vertex.

Errors

Returns an error if the geometry could not be computed.

Examples

Constructing a “spikey” sphere:

use decorum::R64;
use nalgebra::Point3;
use plexus::graph::MeshGraph;
use plexus::prelude::*;
use plexus::primitive::generate::Position;
use plexus::primitive::sphere::UvSphere;

type E3 = Point3<R64>;

let mut graph: MeshGraph<E3> = UvSphere::new(16, 8).polygons::<Position<E3>>().collect();
let keys = graph.faces().map(|face| face.key()).collect::<Vec<_>>();
for key in keys {
    graph.face_mut(key).unwrap().poke_with_offset(0.5).unwrap();
}

pub fn extrude_with_offset<T>(self, offset: T) -> Result<Self, GraphError> where
T: Into<Scalar<VertexPosition<G>>>,
G: FaceNormal,
G::Vertex: AsPositionMut,
VertexPosition<G>: EuclideanSpace,

Extrudes the face along its normal.

Returns the extruded face.

Errors

Returns an error if the geometry could not be computed.

pub fn extrude_with_translation(
self,
translation: Vector<VertexPosition<G>>
) -> Self where
G::Vertex: AsPositionMut,
VertexPosition<G>: EuclideanSpace,

Extrudes the face along a translation.

Returns the extruded face.

pub fn extrude_with<F>(self, f: F) -> Self where
F: Fn(&G::Vertex) -> G::Vertex,

Extrudes a face using the given vertex data.

Returns the extruded face.

pub fn remove(self) -> Option<Ring<&'a mut M>>

Removes the face.

Returns the remaining ring of the face if it is not entirely disjoint, otherwise None.