Trait plexus::index::GroupedIndexVertices

pub trait GroupedIndexVertices<R, P>: Sized where
    R: Grouping,
    P: Topological, {
    fn index_vertices_with<N, K, F>(
        self, 
        indexer: N, 
        f: F
    ) -> (Vec<R::Group>, Vec<P::Vertex>)
    where
        N: Indexer<P, K>,
        F: Fn(&P::Vertex) -> &K;

    fn index_vertices<N>(self, indexer: N) -> (Vec<R::Group>, Vec<P::Vertex>)
    where
        N: Indexer<P, P::Vertex>,
    { ... }
}

Functions for collecting an iterator of $n$-gons into raw index and vertex buffers.

Unlike IndexVertices, this trait provides functions that are closed (not parameterized) with respect to Grouping. Instead, the trait is implemented for a particular Grouping. These functions cannot be used fluently as part of an iterator expression.

Required Methods

fn index_vertices_with<N, K, F>(
self,
indexer: N,
f: F
) -> (Vec<R::Group>, Vec<P::Vertex>) where
N: Indexer<P, K>,
F: Fn(&P::Vertex) -> &K,

Provided Methods

fn index_vertices<N>(self, indexer: N) -> (Vec<R::Group>, Vec<P::Vertex>) where
N: Indexer<P, P::Vertex>,