Module plexus::prelude

Re-exports of commonly used types and traits.

Importing the contents of this module is recommended, especially when working with generators and iterator expressions, as those operations are expressed mostly through traits.

Traits

Traits from the primitive module for generating and decomposing iterators of topological data (e.g., Trigon, Tetragon, etc.) are re-exported so that functions in iterator expressions can be used more easily.

Traits for (de)constructing MeshBuffers and MeshGraphs are re-exported. These traits allow mesh types to be constructed from raw buffers and buffers to be re-indexed.

Extension traits are also re-exported.

Types

The Selector enum and its variants are re-exported for convenience.

Re-exports

pub use crate::buffer::FromRawBuffers as _;

pub use crate::buffer::FromRawBuffersWithArity as _;

pub use crate::buffer::IntoFlatIndex as _;

pub use crate::buffer::IntoStructuredIndex as _;

pub use crate::builder::FacetBuilder as _;

pub use crate::builder::MeshBuilder as _;

pub use crate::builder::SurfaceBuilder as _;

pub use crate::geometry::FromGeometry as _;

pub use crate::geometry::IntoGeometry as _;

pub use crate::graph::Selector;

pub use crate::index::CollectWithIndexer as _;

pub use crate::index::IndexVertices as _;

pub use crate::primitive::decompose::Edges as _;

pub use crate::primitive::decompose::IntoEdges as _;

pub use crate::primitive::decompose::IntoSubdivisions as _;

pub use crate::primitive::decompose::IntoTetrahedrons as _;

pub use crate::primitive::decompose::IntoTrigons as _;

pub use crate::primitive::decompose::IntoVertices as _;

pub use crate::primitive::decompose::Subdivide as _;

pub use crate::primitive::decompose::Tetrahedrons as _;

pub use crate::primitive::decompose::Triangulate as _;

pub use crate::primitive::decompose::Vertices as _;

pub use crate::primitive::generate::Generator as _;

pub use crate::primitive::IntoPolygons as _;

pub use crate::primitive::MapVertices as _;

pub use crate::primitive::Polygonal as _;

pub use crate::primitive::Topological as _;

pub use crate::DynamicArity as _;

pub use crate::IteratorExt as _;

pub use Selector::ByIndex;

pub use Selector::ByKey;