Module plexus::primitive::generate

Polytope generation.

This module provides a generic iterator and traits for generating polygons and vertices containing geometric attributes of polytopes like cubes and spheres. The Generate iterator can be used in iterator expressions.

The primary API of this module is provided by the Generator trait, which is implemented by polytope types like Cube and UvSphere.

Examples

Generating raw buffers from the positional data of a $uv$-sphere:

use nalgebra::Point3;
use plexus::prelude::*;
use plexus::primitive::generate::Position;
use plexus::primitive::sphere::UvSphere;

let sphere = UvSphere::new(16, 16);

// Generate the unique set of positional vertices.
let positions = sphere
    .vertices::<Position<Point3<f64>>>()
    .collect::<Vec<_>>();

// Generate polygons that index the unique set of positional vertices. The
// polygons are decomposed into triangles and then into vertices, where each
// vertex is an index into the position data.
let indices = sphere
    .indexing_polygons::<Position>()
    .triangulate()
    .vertices()
    .collect::<Vec<_>>();

Structs

Generate

Iterator that generates topology and geometric attributes.

Normal

Meta-attribute for surface normals.

Position

Meta-attribute for positions.

Traits

Attribute

Geometric attribute.

AttributeGenerator

AttributePolygonGenerator

AttributeVertexGenerator

Generator

Functions for iterating over the topology and geometry of polytopes.

IndexingPolygonGenerator

PolygonGenerator